Day30/chat-with-pdf/requirements.txt
streamlit
pypdf
requests
python-dotenv
huggingface_hub